+++ src/Galley/LogicalOptimizer/LogicalOptimizer.jl
@@ -71,7 +71,7 @@ function enumerate_distributed_plans(q::PlanNode, visited_plans, alias_hash, max
     end
     return plans
 end
-function high_level_optimize(faq_optimizer::FAQ_OPTIMIZERS, input_plan::PlanNode, ST, alias_stats::Dict{IndexExpr, TensorStats}, alias_hash::Dict{IndexExpr, UInt64}, verbose)
+function high_level_optimize(faq_optimizer::FAQ_OPTIMIZERS, input_plan::PlanNode, ST, alias_stats::Dict{IndexExpr, TensorStats}, alias_hash::Dict{IndexExpr, UInt}, verbose)
     logical_plan = PlanNode[]
     for input_query in input_plan.queries
         logical_queries = high_level_optimize_query(faq_optimizer, input_query, ST, alias_stats, alias_hash, verbose)
@@ -79,12 +79,12 @@ function high_level_optimize(faq_optimizer::FAQ_OPTIMIZERS, input_plan::PlanNode
             alias_hash[query.name.name] = cannonical_hash(query.expr, alias_hash)
             alias_stats[query.name.name] = query.expr.stats
         end
-        append!(logical_plan, logical_queries)    
+        append!(logical_plan, logical_queries)
     end
     return Plan(logical_plan...)
 end
 
-function high_level_optimize_query(faq_optimizer::FAQ_OPTIMIZERS, q::PlanNode, ST, alias_stats::Dict{IndexExpr, TensorStats}, alias_hash::Dict{IndexExpr, UInt64}, verbose)
+function high_level_optimize_query(faq_optimizer::FAQ_OPTIMIZERS, q::PlanNode, ST, alias_stats::Dict{IndexExpr, TensorStats}, alias_hash::Dict{IndexExpr, UInt}, verbose)
     insert_statistics!(ST, q; bindings = alias_stats)
     if faq_optimizer === naive
         insert_node_ids!(q)
@@ -96,7 +96,7 @@ function high_level_optimize_query(faq_optimizer::FAQ_OPTIMIZERS, q::PlanNode, S
     check_dnf = !allequal([n.op.val for n in PostOrderDFS(q) if n.kind === MapJoin])
     q_non_dnf = canonicalize(plan_copy(q), false)
     input_aq = AnnotatedQuery(q_non_dnf, ST)
-    logical_queries, cnf_cost, cost_cache = high_level_optimize_annotated_query(faq_optimizer, input_aq, alias_hash, Dict{UInt64, Float64}(), verbose)
+    logical_queries, cnf_cost, cost_cache = high_level_optimize_annotated_query(faq_optimizer, input_aq, alias_hash, Dict{UInt, Float64}(), verbose)
     if check_dnf
         min_cost = cnf_cost
         min_query = canonicalize(plan_copy(q), false)
@@ -133,7 +133,7 @@ function high_level_optimize_query(faq_optimizer::FAQ_OPTIMIZERS, q::PlanNode, S
     return logical_queries
 end
 
-function high_level_optimize_annotated_query(faq_optimizer::FAQ_OPTIMIZERS, aq::AnnotatedQuery, alias_hash::Dict{IndexExpr, UInt64}, cost_cache::Dict{UInt64, Float64}, verbose)
+function high_level_optimize_annotated_query(faq_optimizer::FAQ_OPTIMIZERS, aq::AnnotatedQuery, alias_hash::Dict{IndexExpr, UInt}, cost_cache::Dict{UInt, Float64}, verbose)
     if faq_optimizer == greedy
         return pruned_query_to_plan(aq, cost_cache, alias_hash; use_greedy=true)
     elseif faq_optimizer == exact
